Update and rename CS112_A1_T3_20230779_20230780_20230783.py to Six Problems.py
The team will work together on solving these problems. They should divide the work equally.
Solve every problem in a separate function with a menu that allows the user to choose which
problem to solve and repeats the choice until the user exists.
1- Write a Python program that takes the mark and prints the grade. A+ ≥ 90, 90 &gt; A ≥ 85, etc.
2- Write a Python program to check whether a given number is an Armstrong number or not. A
positive integer of 3 digits is called an Armstrong number if the sum of the cubes of
individual digits of the number is equal to that number itself. For example, the sum of cubes
of individual digits of the number 153 is 1* 1 * 1 + 5 * 5 * 5 + 3 * 3 * 3 = 153. Hence,
number 153 is an Armstrong number. 9474 is an Armstrong number because the sum of each
digit to the power of number size of 4 is 9 4 + 4 4 + 7 4 + 4 4 = 9474.
3- The German mathematician Leibniz (1646–1716) discovered the rather remarkable fact that
the mathematical constant () Pi ط can be computed using the following mathematical
relationship:

The formula to the right of the equal sign represents an infinite series; each fraction
represents a term in that series. If you start with 1, subtract one-third, add one-fifth, and so
on, for each of the odd integers, you get a number that gets closer and closer to the value of
/4 as you go along. Then we can get .
a. Write a detailed algorithm in English or pseudo-code that solves this problem by
calculating n terms, which the user inputs and then calculating . Implement it in
Python.
b. Test your program for n = 10, i.e., calculate 1-1/3+1/5-1/7+1/9-1/11+1/13-1/15+1/17-
1/19

4- Encryption is a very very important topic in computing. It means changing the data so that
enemies (or bad people) do not know what it is. This is needed, for example, when
transferring financial data like bank account numbers and credit card numbers through the
internet.
Write an algorithm (in pseudo-code) and Python program that takes a message from the user
and converts it to an encrypted version using Cesar cipher. Cesar cipher, in the simplest
form, is shifting each character to the next one. So the message &quot;abc&quot; becomes &quot;bcd&quot; and the
message &quot;I love computers&quot; become &quot;J mpwf dpnqvufst&quot;. Hint: Think of a solution that uses
the ASCII code of the letter to generate the replacement letter.
5- Develop a Python function that checks if two lists have the same elements, even if not in
order. Do not use built in functions.
Sample input: [1, 2, 3, 4, 5, 6, 7, 8] and [8, 7, 6, 5, 4, 3, 2, 1]
Output: Lists are equal = True
Sample input: [1, 2, 3, 4, 44, 6, 7, 8] and [8, 7, 6, 5, 4, 3, 2, 1]
Output: Lists are equal = False
6- Design an algorithm and Python program to find all the factors of a +ve integer. For example,
in the case of the integer 12, your algorithm should report the values 1, 2, 3, 4, 6, and 12.