Create Cesar Cipher Encryption.py
Encryption is a very very important topic in computing. It means changing the data so that
enemies (or bad people) do not know what it is. This is needed, for example, when
transferring financial data like bank account numbers and credit card numbers through the
internet.
Write an algorithm (in pseudo-code) and Python program that takes a message from the user
and converts it to an encrypted version using Cesar cipher. Cesar cipher, in the simplest
form, is shifting each character to the next one. So the message "abc" becomes "bcd" and the
message "I love computers" become "J mpwf dpnqvufst".